+++ test/unit/test_arguments.py
@@ -24,8 +24,16 @@ from itertools import zip_longest
 
 
 def test_simple_dict():
-    dict_str = 'beta1:0.9,beta2:0.999,epsilon:1e-8,lazy_update:true'
-    expected = {'beta1': 0.9, 'beta2': 0.999, 'epsilon': 1e-8, 'lazy_update': True}
+    dict_str = 'a:True,b:true,' \
+               'c:False,d.e:FALSE,' \
+               'f.g:1,h.i:-234,' \
+               'j.k.l:1.,m.n.o:.1,p:1e-10,' \
+               'q:`~!@#$%^&*()-_=+[{]}\\|;\'"<.>/?,r:str'
+    expected = {'a': True, 'b': True,
+                'c': False, 'd.e': False,
+                'f.g': 1, 'h.i': -234,
+                'j.k.l': 1., 'm.n.o': .1, 'p': 1e-10,
+                'q': '`~!@#$%^&*()-_=+[{]}\\|;\'"<.>/?', 'r': 'str'}
     parse = arguments.simple_dict()
     assert parse(dict_str) == expected
 
@@ -218,7 +226,6 @@ def test_inference_args(test_params, expected_params):
               gradient_clipping_threshold=1.0,
               gradient_clipping_type='none',
               learning_rate_scheduler_type='plateau-reduce',
-              learning_rate_t_scale=1.0,
               learning_rate_reduce_factor=0.9,
               learning_rate_reduce_num_not_improved=8,
               learning_rate_warmup=0,