+++ CHANGELOG.md
@@ -11,6 +11,27 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [3.1.19]
+
+### Added
+
+- Clean up GPU and CPU memory used during training initialization before starting the main training loop.
+
+### Changed
+
+- Refactored training code in advance of adding DeepSpeed support:
+  - Moved logic for flagging interleaved key-value parameters from layers.py to model.py.
+  - Refactored LearningRateScheduler API to be compatible with PyTorch/DeepSpeed.
+  - Refactored optimizer and learning rate scheduler creation to be modular.
+  - Migrated to ModelWithLoss API, which wraps a Sockeye model and its losses in a single module.
+  - Refactored primary and secondary worker logic to reduce redundant calculations.
+  - Refactored code for saving/loading training states.
+  - Added utility code for managing model/training configurations.
+
+### Removed
+
+- Removed unused training option `--learning-rate-t-scale`.
+
 ## [3.1.18]
 
 ### Added